btcec: Make it easier to build with/without libsecp256k1
