--- sipasec/.gitignore
@@ -1,1 +1,0 @@
-secp256k1.h
