--- sipasec/sipasec_unix.go
@@ -1,20 +1,0 @@
-// +build darwin dragonfly freebsd linux openbsd solaris netbsd
-
-package sipasec
-
-/*
-To build and install secp256k1 lib on Debian Linux system, execute the following steps:
-
- * sudo apt-get install gcc autoconf libtool make
- * git clone https://github.com/bitcoin/bitcoin.git
- * cd bitcoin/src/secp256k1/
- * ./autogen.sh
- * ./configure
- * make
- * sudo make install
-
-When the lib is properly installed, executing "go test" in this folder will say "PASS".
-Then copy "gocoin/client/speedups/sipasec.go" to "gocoin/client/" to boost your client.
-*/
-
-import "C"
